refactor: review: move types out of const file to its own file
