+++ src/plugins/duplication-detection/index.test.ts
@@ -6,7 +6,7 @@ import DuplicationDetectionPlugin from './index';
 import baseCompilationStats from '../../../test/fixtures/base-compilation-stats.json';
 import headCompilationStats from '../../../test/fixtures/head-compilation-stats.json';
 import { defaultDuplicationDetectionConfig } from './config';
-import { Stats4 } from '../../helpers/constants';
+import { Stats4 } from '../../types';
 import extractStats from '../../helpers/extractStats';
 
 const baseStats = (baseCompilationStats as unknown) as Stats4;
