+++ src/plugins/trace-changes/index.test.ts
@@ -5,7 +5,7 @@
 import TraceChangesPlugin from './index';
 import baseCompilationStats from '../../../test/fixtures/base-compilation-stats.json';
 import headCompilationStats from '../../../test/fixtures/head-compilation-stats.json';
-import { Stats4 } from '../../helpers/constants';
+import { Stats4 } from '../../types';
 import extractStats from '../../helpers/extractStats';
 
 const baseStats = (baseCompilationStats as unknown) as Stats4;
